requirements.txt
Django==4.2.1
psycopg2-binary==2.9.6
python-dotenv==1.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.5.0